Add test cases for example files under test/
This adds a new test case which will try to parse each one of the files
at "test/*.conf". So far the corpus of examples includes the example
files from libwheel's wcfg module, and some configuration examples from
the Drove monitorization system. The files are included with permission
of their respective authors.