Add REST web service example converter example
This adds an examples/ subdirectory, and an example (using Bottle) which shows
how to make a simple web application which accepts REST requests with POST
data in different formats, and returns the equivalent HiPack message. This is
to be used in http://hipack.org, so it handles CORS requests from that domain.