HEP-1: Support on-parse value cast callbacks
This wires the parsing of annotations with the rest of the code in order to
allow doing useful things with the annotations. By default, behaviour will be
as before, but it is also now possible to specify a callback function which
converts parsed values on-the-go, and has access to annotations.

Issue: #3