Fix error scanning hex escape sequences in strings
When converting an hex escape code inside a string literal, in Python 3 the
resulting value would be converted to a "str", while the rest of the parser
expects input as "bytes". This patch makes sure that an hex escape code
sequence is always converted to "bytes" before going back to the parser code
path.